requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
dvc==3.30.0
pytest==7.4.0
flake8==6.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
apache-airflow==2.7.3
requests==2.31.0
python-multipart==0.0.6
